+++ changelog.md
@@ -1,9 +1,10 @@
-1.27.1
-===================
+Upcoming
+========
 
 Bug Fixes:
----------
+----------
 
+* Don't install tests.
 * Fix unexpected exception when using dsn without username & password (Thanks: [Will Wang])
 
 1.27.0 (2023/08/11)
@@ -22,6 +23,7 @@ Bug Fixes:
 * Remove vi-mode bindings for prettify/unprettify.
 * Honor `\G` when executing from commandline with `-e`.
 * Correctly report the version of TiDB.
+* Revised `botton` spelling mistakes with `bottom` in `mycli/clitoolbar.py`
 
 
 1.26.1 (2022/09/01)